backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
boto3==1.34.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]==2.5.0
pymysql==1.1.0
bcrypt==4.0.1